refactor(dashboard): extract badges + sse sub-routers (F-B-202 PR-4/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-4 of 10. Extracts 7 routes (6 HTMX badges
+ 1 SSE) from the router god-object. Originally planned to also
extract the overview landing page route, but FastAPI rejects
sub-router routes with an empty path (the bare ``/dashboard``
endpoint uses ``@router.get("")``); the overview handler stays
inline with a footnote comment explaining the constraint.

What moves (7 routes, ~155 LOC):

  app/dashboard/badges_routes.py (new, 98 LOC, 6 routes):
    GET /dashboard/badge/pending-orgs
    GET /dashboard/badge/pending-sessions
    GET /dashboard/badge/users-count
    GET /dashboard/badge/agents-count
    GET /dashboard/badge/workloads-count
    GET /dashboard/badge/resources-count

  app/dashboard/sse_routes.py (new, 53 LOC, 1 route):
    GET /dashboard/sse

  app/dashboard/router.py (2614 → 2372 LOC, -242):
    - 7 routes removed; overview kept inline (FastAPI empty-path
      constraint on sub-routers).
    - Two new include_router calls near the top.

Behavioural delta: zero. URLs identical, response bodies identical,
SSE event stream semantics unchanged.

Pre-merge gate per memory feedback_legacy_sandbox_smoke_obsolete:
  - Full unit suite: 4216 passed, 9 skipped, 32 xfailed in ~115s
    (1 xdist flake test_ensure_private_key_permissions_warns_on_loose_mode
    pre-existing, passes isolated)
  - Dashboard targeted: pytest tests/test_dashboard*.py → 202/203
    (1 pre-existing deselect)
  - Stack smoke ./stack/smoke.sh NOT run for this PR: the modern
    stack smoke exercises Mastio + Connector + Frontdesk paths, not
    the Court /dashboard surface this refactor touches. Refactor
    is mechanical (extract + include_router, same pattern as
    PR-1/PR-2/PR-3), so unit suite is the load-bearing gate.

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC
  Post-PR-2 #841: 60 routes, 2614 LOC
  Post-PR-3 #842: 57 routes, 2528 LOC
  Post-PR-4 (this): 50 routes, 2372 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Next PR-5 (F-B-202 modularization): extract orgs_routes.py
(CRUD + invites + approve/reject/suspend/delete/upload-CA, ~350 LOC,
11 routes — biggest single-PR extract so far).
